refactor: move test prep to common location
