+++ code/system/sys.fs/src/TEST/TestPrep.mts
@@ -1,0 +1,21 @@
+import { t, slug, rx } from '../common/index.mjs';
+import { FsMock } from '../Mock/index.mjs';
+import { BusController } from '../BusController/index.mjs';
+
+/**
+ * Setup a mock FS driver and Controller.
+ */
+export const TestPrep = (options: { dir?: string; id?: string } = {}) => {
+  const { dir, id = `foo.${slug()}` } = options;
+
+  const driver = FsMock.Driver({ dir }).driver;
+  const indexer = FsMock.Indexer({ dir }).indexer;
+
+  const bus = rx.bus<t.SysFsEvent>();
+  const controller = BusController({ id, driver, bus, indexer });
+
+  const events = controller.events;
+  const { dispose } = events;
+
+  return { bus, controller, events, dispose, driver, indexer };
+};
