+++ code/system/sys.fs/src/TEST/index.mts
@@ -1,4 +1,6 @@
-export { expect, expectError } from 'sys.test';
 export { beforeEach, describe, it } from 'vitest';
-export * from '../Mock/index.mjs';
+export { expect, expectError } from 'sys.test';
+
 export * from '../common/index.mjs';
+export { FsMock } from '../Mock/index.mjs';
+export { TestPrep } from './TestPrep.mjs';
